+++ packages/cli/src/bin.ts
@@ -1,5 +1,5 @@
 #!/usr/bin/env node
 import { runMain } from "citty";
-import main from "./index";
+import main from "./index.js";
 
 runMain(main);